requirements.txt
beautifulsoup4==4.9.3
requests==2.25.1
python-dotenv==0.19.2
python-wordpress-xmlrpc==2.3
lxml==4.6.3
